pyproject.toml
[project]
name = "aha-moments"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"

# Remember to download CUDA toolkit here: 
# https://developer.nvidia.com/cuda-13-0-0-download-archive?target_os=Windows&target_arch=x86_64&target_version=11&target_type=exe_local

dependencies = [
    "cupy-cuda13x[all]>=13.6.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
]
